Add Weapon migration and changes on crud characters / user

+++ Services/CharacterService/CharacterService.cs
@@ -1,11 +1,13 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 using AutoMapper;
 using Dot_Net_Core_API_with_JWT.Data;
 using Dot_Net_Core_API_with_JWT.Dtos.Character;
 using Dot_Net_Core_API_with_JWT.Models;
+using Microsoft.AspNetCore.Http;
 using Microsoft.EntityFrameworkCore;
 
 namespace Dot_Net_Core_API_with_JWT.Services.CharacterService
@@ -14,20 +16,28 @@ namespace Dot_Net_Core_API_with_JWT.Services.CharacterService
   {
     private readonly IMapper _mapper;
     private readonly DataContext _context;
+    private readonly IHttpContextAccessor _httpContextAccessor;
 
-    public CharacterService(IMapper mapper, DataContext context)
+    public CharacterService(IMapper mapper, DataContext context, IHttpContextAccessor httpContextAccessor)
     {
+      _httpContextAccessor = httpContextAccessor;
       _mapper = mapper;
       _context = context;
     }
+
+    private int GetUserId() => int.Parse(_httpContextAccessor.HttpContext.User.FindFirstValue(ClaimTypes.NameIdentifier));
 
     public async Task<ServiceResponse<List<GetCharacterDto>>> AddCharacter(AddCharacterDto newCharacter)
     {
       var serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
       Character character = _mapper.Map<Character>(newCharacter);
+      character.User = await _context.Users.FirstOrDefaultAsync(u => u.Id == GetUserId());
+
       _context.Characters.Add(character); //Add character
       await _context.SaveChangesAsync();
-      serviceResponse.Data = await _context.Characters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToListAsync();
+      serviceResponse.Data = await _context.Characters
+      .Where(c => c.User.Id == GetUserId()) //Retorna todos o characteres do usuário logado
+      .Select(c => _mapper.Map<GetCharacterDto>(c)).ToListAsync();
       return serviceResponse;
     }
 
@@ -37,10 +47,20 @@ namespace Dot_Net_Core_API_with_JWT.Services.CharacterService
 
       try
       {
-        Character character = await _context.Characters.FirstAsync(c => c.Id == id);
-        _context.Characters.Remove(character);
-        await _context.SaveChangesAsync();
-        serviceResponse.Data = _context.Characters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
+        Character character = await _context.Characters.FirstOrDefaultAsync(c => c.Id == id && c.User.Id == GetUserId());
+        if (character != null)
+        {
+          _context.Characters.Remove(character);
+          await _context.SaveChangesAsync();
+          serviceResponse.Data = _context.Characters
+            .Where(c => c.User.Id == GetUserId())
+            .Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
+        }
+        else
+        {
+          serviceResponse.Success = false;
+          serviceResponse.Message = "Personagem não encontrado.";
+        }
       }
       catch (Exception ex)
       {
@@ -50,11 +70,11 @@ namespace Dot_Net_Core_API_with_JWT.Services.CharacterService
       return serviceResponse;
     }
 
-    public async Task<ServiceResponse<List<GetCharacterDto>>> GetAllCharacters(int userId)
+    public async Task<ServiceResponse<List<GetCharacterDto>>> GetAllCharacters()
     {
       var serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
 
-      var dbCharacters = await _context.Characters.Where(c => c.User.Id == userId).ToListAsync();
+      var dbCharacters = await _context.Characters.Where(c => c.User.Id == GetUserId()).ToListAsync();
       serviceResponse.Data = dbCharacters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
       return serviceResponse;
     }
@@ -62,7 +82,7 @@ namespace Dot_Net_Core_API_with_JWT.Services.CharacterService
     public async Task<ServiceResponse<GetCharacterDto>> GetCharacterById(int id)
     {
       var serviceResponse = new ServiceResponse<GetCharacterDto>();
-      var dbCharacter = await _context.Characters.FirstOrDefaultAsync(c => c.Id == id);
+      var dbCharacter = await _context.Characters.FirstOrDefaultAsync(c => c.Id == id && c.User.Id == GetUserId());
       serviceResponse.Data = _mapper.Map<GetCharacterDto>(dbCharacter);
       return serviceResponse;
     }
@@ -73,18 +93,25 @@ namespace Dot_Net_Core_API_with_JWT.Services.CharacterService
 
       try
       {
-        Character character = await _context.Characters.FirstOrDefaultAsync(c => c.Id == updatedCharacter.Id);
+        Character character = await _context.Characters
+          .Include(c => c.User)
+          .FirstOrDefaultAsync(c => c.Id == updatedCharacter.Id);
+        if(character.User.Id == GetUserId())
+        {
+          character.Name = updatedCharacter.Name;
+          character.HitPoints = updatedCharacter.HitPoints;
+          character.Strength = updatedCharacter.Strength;
+          character.Defense = updatedCharacter.Defense;
+          character.Intelligence = updatedCharacter.Intelligence;
+          character.Class = updatedCharacter.Class;
+          await _context.SaveChangesAsync();
+          serviceResponse.Data = _mapper.Map<GetCharacterDto>(character);
+        }
+        else{
+          serviceResponse.Success = false;
+          serviceResponse.Message = "Personagem não encontrado.";
+        }
 
-        character.Name = updatedCharacter.Name;
-        character.HitPoints = updatedCharacter.HitPoints;
-        character.Strength = updatedCharacter.Strength;
-        character.Defense = updatedCharacter.Defense;
-        character.Intelligence = updatedCharacter.Intelligence;
-        character.Class = updatedCharacter.Class;
-
-        await _context.SaveChangesAsync();
-
-        serviceResponse.Data = _mapper.Map<GetCharacterDto>(character);
       }
       catch (Exception ex)
       {
